+++ pyobjc-framework-PushKit/Lib/PushKit/__init__.py
@@ -1,0 +1,26 @@
+'''
+Python mapping for the PushKit framework.
+
+This module does not contain docstrings for the wrapped code, check Apple's
+documentation for details on how to use these functions and classes.
+'''
+
+import objc
+import sys
+import Foundation
+
+from PushKit import _metadata
+
+sys.modules['PushKit'] = mod = objc.ObjCLazyModule(
+    "PushKit",
+    "com.apple.pushkit",
+    objc.pathForFramework("/System/Library/Frameworks/PushKit.framework"),
+    _metadata.__dict__, _inline_list_, {
+        '__doc__': __doc__,
+        'objc': objc,
+        '__path__': __path__,
+        '__loader__': globals().get('__loader__', None),
+    }, (Foundation,))
+
+import sys
+del sys.modules['PushKit._metadata']